# api/reference/openapi.yaml
openapi: 3.0.3
info:
  title: Timeterm
  version: "1.0"
  contact:
    name: Timeterm
    email: support@timeterm.nl
  description: The Timeterm API
security:
  - ApiKeyAuth: []
servers:
  - url: "https://api.timeterm.nl/"
paths:
...
  /device:
    post:
      operationId: registerDevice
      summary: Register a new device
      description: |
        Register a new device

        > NOTE: this API call can only be performed by devices.
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Device"
      responses:
        "200":
          description: Device created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CreateDeviceResponse"
        default:
          $ref: "#/components/responses/ProblemResponse"
    get:
      operationId: getDevices
      summary: List devices
      parameters:
        - name: offset
          in: query
          schema:
            type: integer
            format: uint64
        - name: maxAmount
          in: query
          schema:
            type: integer
            format: uint64
        - name: searchName
          in: query
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaginatedDevices"
    delete:
      operationId: deleteDevices
      summary: Delete devices
      requestBody:
        content:
          application/json:
            schema:
              properties:
                deviceIds:
                  type: array
                  items:
                    type: string
                    format: uuid
      responses:
        "204":
          description: No content
  /device/restart:
    post:
      operationId: rebootDevices
      summary: Reboot a list of devices
      requestBody:
        content:
          application/json:
            schema:
              properties:
                deviceIds:
                  type: array
                  items:
                    type: string
                    format: uuid
      responses:
        "200":
          description: No content
...
components:
  schemas:
...
    Device:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        organizationId:
          type: string
          format: uuid
        primaryStatus:
          $ref: "#/components/schemas/PrimaryDeviceStatus"
        secondaryStatus:
          $ref: "#/components/schemas/SecondaryDeviceStatus"
    PrimaryDeviceStatus:
      type: string
      enum:
        - Online
        - Offline
    SecondaryDeviceStatus:
      type: string
      enum:
        - NotActivated
...
    Pagination:
      properties:
        offset:
          type: integer
          format: uint64
        maxAmount:
          type: integer
          format: uint64
        total:
          type: integer
          format: uint64
    PaginatedDevices:
      allOf:
        - $ref: "#/components/schemas/Pagination"
        - properties:
            data:
              type: array
              items:
                $ref: "#/components/schemas/Device"
...
    Problem:
      description: An RFC7807 Problem Details Object.
      properties:
        type:
          description: A URI reference [[RFC 3986](https://tools.ietf.org/html/rfc3986)] that identifies the problem type. If empty, assumed to be "about:blank".
          type: string
        title:
          type: string
          description: A short, human-readable summary of the problem type.
        status:
          type: integer
          description: The HTTP status code ([[RFC 7231](https://tools.ietf.org/html/rfc7231#section-6)], Section 6) generated by the origin server for this occurrence of the problem.
        detail:
          type: string
          description: A human-readable explanation specific to this occurrence of the problem.
        instance:
          type: string
          description: A URI reference that identifies the specific occurrence of the problem. It may or may not yield further information if dereferenced.
    CreateDeviceResponse:
      properties:
        device:
          $ref: "#/components/schemas/Device"
        token:
          type: string
...
  responses:
    ProblemResponse:
      description: An RFC 7807 Problem Details Object response.
      content:
        "application/json":
          schema:
            $ref: "#/components/schemas/Problem"
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-Api-Key